Module for webscrappers.
