--- src/main/java/com/java/academy/model/bookstore/googlebookstore/ImageLinks.java
@@ -1,25 +1,0 @@
-package com.java.academy.model.bookstore.googlebookstore;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-/**
- * @author bratek
- * @since 24.08.17
- */
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ImageLinks {
-    private String smallThumbnail;
-
-    public ImageLinks() {
-    }
-
-    public String getSmallThumbnail() {
-        return smallThumbnail;
-    }
-
-    public void setSmallThumbnail(String smallThumbnail) {
-        this.smallThumbnail = smallThumbnail;
-    }
-
-
-}
