--- src/main/java/com/java/academy/model/bookstore/googlebookstore/RetailPrice.java
@@ -1,24 +1,0 @@
-package com.java.academy.model.bookstore.googlebookstore;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-/**
- * @author bratek
- * @since 24.08.17
- */
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class RetailPrice {
-    private Double amount;
-
-    public RetailPrice() {
-    }
-
-    public Double getAmount() {
-        return amount;
-    }
-
-    public void setAmount(Double amount) {
-        this.amount = amount;
-    }
-
-}
