--- webscrapper/src/main/java/com/java/academy/WebscrapperApplication.java
@@ -1,12 +1,0 @@
-package com.java.academy;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WebscrapperApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(WebscrapperApplication.class, args);
-	}
-}
