--- webscrapper/src/test/java/com/java/academy/WebscrapperApplicationTests.java
@@ -1,16 +1,0 @@
-package com.java.academy;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class WebscrapperApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
